complete refactor of index.ts SRP initialized
