+++ src/routes/health.ts
@@ -1,0 +1,15 @@
+import { Router } from "express";
+import { prisma } from "../lib/prisma";
+
+export const healthRouter = Router();
+
+healthRouter.get("/health", async (_req, res) => {
+  try {
+    await prisma.$queryRaw`SELECT 1`;
+    res.json({ ok: true });
+  } catch (err) {
+    console.error(err);
+    res.status(500).json({ ok: false });
+  }
+});
+
